collapsed check in/check out menu for employees
